requirements.txt
Django==1.11
djangorestframework==3.9.0
pkg-resources==0.0.0
pytz==2018.7
